#10 [refactor] TierList 분리
